requirements.txt
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
geopy>=2.3.0
folium>=0.14.0
pulp>=2.7.0
pytest>=7.4.0
python-dateutil>=2.8.0
tqdm>=4.65.0
